requirements-train.txt
torch==2.3.1
transformers==4.40.2
datasets==2.20.0
sentence-transformers==3.0.1
accelerate==0.31.0
bitsandbytes==0.43.1
peft==0.11.1
pyyaml
pydantic>=2.6.0
huggingface_hub==0.23.4
dspy-ai
